pyproject.toml
[tool.poetry]
name = "nova-fastapi-microservice"
version = "0.1.0"
description = "design template for building microservice using fastapi"
authors = ["Michael Asumadu <masumadu@archsolutions.online>"]
readme = "README.md"
packages = [{include = "nova_fastapi_microservice"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.110.0"
pre-commit = "^3.2.2"
sqlalchemy = "^2.0.9"
psycopg2-binary = "^2.9.6"
pinject = "^0.14.1"
pyjwt = "^2.6.0"
redis = "^4.5.4"
kafka-python = "^2.0.2"
alembic = "^1.10.3"
loguru = "^0.7.0"
fastapi-pagination = "^0.12.2"
isort = "^5.12.0"
blinker = "^1.6.2"
pytz = "^2023.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptography = "43.0.3"
requests = "2.32.3"
jinja2 = "3.1.4"
pybreaker = "^1.0.2"
tenacity = "^8.2.2"
flower = "^2.0.0"
fastapi-limiter = "^0.1.5"
aioredis = "^2.0.1"
fakeredis = {extras = ["lua"], version = "^2.18.0"}
pydantic = {extras = ["email"], version = "^2.10.3"}
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
httpx = "^0.25.0"
smpplib = "^2.2.3"
faker = "^24.8.0"
kombu = "5.3.7"
celery = "5.4.0"
anyio = "4.4.0"
idna = "3.7"
urllib3 = "2.2.2"
tornado = "6.4.1"
certifi = "2024.7.4"
openpyxl = "^3.1.5"
setuptools = "75.2.0"
virtualenv = "20.27.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "24.4.0"
pytest = "^7.3.0"
coverage = "^7.2.3"
pytest-mock = "^3.10.0"
safety = "^2.3.5"
autoflake = "^2.2.1"
isort = "^5.12.0"

[tool.black]
line-length = 89
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(/(
    \.git
  | \.hg
  | \.eggs
  | \.ini
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | .yml
  | .yaml
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
)/
)
'''

[[tool.poetry.source]]
name = "masumadu"
url = "https://test.pypi.org/simple/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
